Autofocus: stop exposing internal AutofocusManager in API.
Instead we have a new AutofocusManager interface that only exposes specific
methods, which our old AutofocusManager (renamed to DefaultAutofocusManager)
implements.

Along the way, tweaked some method names.
