+++ mmstudio/src/main/java/org/micromanager/Studio.java
@@ -28,7 +28,6 @@ import mmcorej.CMMCore;
 import org.micromanager.data.DataManager;
 import org.micromanager.display.DisplayManager;
 import org.micromanager.events.EventManager;
-import org.micromanager.internal.utils.AutofocusManager;
 import org.micromanager.quickaccess.QuickAccessManager;
 
 
@@ -95,7 +94,8 @@ public interface Studio {
    public AcquisitionManager getAcquisitionManager();
 
    /**
-    * @return the currently selected AutoFocusManger object
+    * Provide access to the AutofocusManager, for performing autofocus
+    * operations.
     */
    public AutofocusManager getAutofocusManager();
